src/requirements.txt
uvicorn[standard]
discord.py
requests
dotenv